requirements.txt
pandas>=3.0.0
python-docx>=1.2.0
openpyxl>=3.0.0
